+++ changelog.md
@@ -10,6 +10,7 @@
 - Captain icon replaces captain title for selected and undrafted captains on Game Detail Page (saves space)
 - caching series games and player counts instead of annotating
 - Added "Loading" above spinner to make the animation more clear
+- Renamed scorecards to box scores
 
 ## [1.12.15] - 2026-7-08 Leaders